refactor: TreeView component
