+++ src/styles.module.css
@@ -1,7 +1,7 @@
 .nodeWrapper {
   align-items: center;
   display: grid;
-  grid-template-columns: auto 1fr auto auto auto;
+  grid-template-columns: auto 1fr auto auto;
   height: 32px;
   padding-inline-end: 8px;
   padding-inline-start: 8px;
